fix: rename tabs-nav
