+++ packages/zent/src/tabs/components/tabs-nav/NormalTabsNav.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { IInnerTab } from '../../types';
 import NormalTab from '../tab/NormalTab';
-import CommonTabsNav from './common';
+import CommonTabsNav from './CommonTabsNav';
 
 class NormalTabsNav<Id extends string | number = string> extends CommonTabsNav<
   Id
